requirements.txt
psycopg2-binary==2.9.3
psycopg2==2.9.3
python-dotenv==0.20.0